requirements.txt
streamlit>=1.24.0
pandas>=1.5.3
requests>=2.28.2
numpy>=1.24.3
tqdm>=4.65.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
loguru==0.7.2 